@@ -23,10 +23,10 @@ def to_element string
   return 'user_email' if string == "the register email"
   return 'user_password' if string == "the register password"
   return 'form#new_user input[type="submit"]' if string == "the register form"
-  return '.field_with_errors label[for="user_first_name"]' if string == "first name field error"
-  return '.field_with_errors label[for="user_last_name"]' if string == "last name field error"
-  return '.field_with_errors label[for="user_email"]' if string == "email field error"
-  return '.field_with_errors label[for="user_password"]' if string == "password field error"
+  return '.first_name.error small.error' if string == "first name field error"
+  return '.last_name.error small.error' if string == "last name field error"
+  return '.email.error small.error' if string == "email field error"
+  return '.password.error small.error' if string == "password field error"
   return '.flash.alert' if string == "an error message"
   return 'user_profession' if string == "the profession field"
   return 'user_website' if string == "the website field"
@@ -37,14 +37,13 @@ def to_element string
   return 'user_gender_male' if string == "male as gender"
   return 'user_postal_code' if string == "the postal code field"
   return 'user_phone' if string == "the phone field"
-  return 'user_phone' if string == "the phone field"
   return 'user_email' if string == "the email field"
   return 'user_secondary_email' if string == "the secondary email field"
-  return 'label[for="user_email"].message' if string == "the email error"
-  return 'label[for="user_secondary_email"].message' if string == "the secondary email error"
+  return '.error label[for="user_email"]' if string == "the email error"
+  return '.error label[for="user_secondary_email"]' if string == "the secondary email error"
   return 'label[for="user_postal_code"].message' if string == "the postal code error"
   return 'label[for="user_phone"].message' if string == "the phone error"
-  return 'label[for="user_website"].message' if string == "the website error"
+  return '.error label[for="user_website"]' if string == "the website error"
   return 'form.edit_user input[type="submit"]' if string == "the edit profile form submition button"
   return '.flash.alert' if string == "you have no authorization to access this page"
   return "form#edit_user_#{@user.id}" if string == "this user form edit"
